pyproject.toml
[project]
name = "atividade2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.2.6",
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
    "streamlit>=1.47.0",
]
